applications/faq_tool/requirements.txt
python-dotenv~=1.1.0
gradio~=5.23.1
openai~=1.69.0
chromadb~=0.6.3
pymupdf4llm~=0.0.17
tqdm~=4.67.1
